refactor: utils class in utils package
